Created base map view controller, moving logic into that
